requirements.txt
blenderproc==2.8.0
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
h5py==3.13.0
idna==3.10
kiwisolver==1.4.8
mathutils==3.3.0
matplotlib==3.10.3
numpy==2.2.5
packaging==25.0
pillow==11.2.1
progressbar==2.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
setuptools==80.4.0
six==1.17.0
urllib3==2.4.0
